move helper to hypertypes
